Corrigiendo doc y estructurando directorios con pruebas.
