+++ src/pruebas/PruebaPixel.cs
@@ -1,0 +1,15 @@
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+using System;
+using ProcesadorImagen.Pixel;
+
+    ///[TestClass]
+    public class PruebaPixel {
+
+        ///[TestMethod]
+        public void PruebaConstructorPixel() {
+            Pixel pixel = new Pixel(1, 2, Color.Red);
+            Assert.AreEqual(1, pixel.ObtenX());
+            Assert.AreEqual(2, pixel.ObtenY());
+            Assert.AreEqual(Color.Red, pixel.ObtenColor());
+        }
+    }
